+++ spy/errors.py
@@ -1,71 +1,7 @@
-from typing import Optional, Literal, ClassVar
+from typing import ClassVar
 from dataclasses import dataclass
-import linecache
 from spy.location import Loc
-from spy.textbuilder import ColorFormatter
-
-Level = Literal["error", "note", "panic"]
-
-def maybe_plural(n: int, singular: str, plural: Optional[str] = None) -> str:
-    if n == 1:
-        return singular
-    elif plural is None:
-        return singular + 's'
-    else:
-        return plural
-
-@dataclass
-class Annotation:
-    level: Level
-    message: str
-    loc: Loc
-
-
-class ErrorFormatter:
-    err: 'SPyError'
-    lines: list[str]
-
-    def __init__(self, err: 'SPyError', use_colors: bool) -> None:
-        self.err = err
-        self.color = ColorFormatter(use_colors)
-        # add "custom colors" to ColorFormatter, so that we can do
-        # self.color.set('error', 'hello')
-        self.color.error = self.color.red  # type: ignore
-        self.color.panic = self.color.red  # type: ignore
-        self.color.note = self.color.green # type: ignore
-        self.lines = []
-
-    def w(self, s: str) -> None:
-        self.lines.append(s)
-
-    def build(self) -> str:
-        return '\n'.join(self.lines)
-
-    def emit_message(self, level: Level, message: str) -> None:
-        prefix = self.color.set(level, level)
-        message = self.color.set('default', message)
-        self.w(f'{prefix}: {message}')
-
-    def emit_annotation(self, ann: Annotation) -> None:
-        filename = ann.loc.filename
-        line = ann.loc.line_start
-        col = ann.loc.col_start + 1  # Loc columns are 0-based but we want 1-based
-        srcline = linecache.getline(filename, line).rstrip('\n')
-        carets = self.make_carets(srcline, ann.loc, ann.message)
-        carets = self.color.set(ann.level, carets)
-        self.w(f'   --> {filename}:{line}:{col}')
-        self.w(f'{line:>3} | {srcline}')
-        self.w(f'    | {carets}')
-        self.w('')
-
-    def make_carets(self, srcline: str, loc: Loc, message: str) -> str:
-        a = loc.col_start
-        b = loc.col_end
-        if b < 0:
-            b = len(srcline) + b + 1
-        n = b-a
-        line = ' ' * a + '^' * n
-        return line + ' ' + message
+from spy.errfmt import ErrorFormatter, Level, Annotation
 
 
 class SPyError(Exception):
