+++ spy/location.py
@@ -110,7 +110,7 @@ class Loc:
         """
         Visualize the piece of code which correspond to this Loc
         """
-        from spy.errors import ErrorFormatter, Annotation
+        from spy.errfmt import ErrorFormatter, Annotation
         fmt = ErrorFormatter(err=None, use_colors=True) # type: ignore
         ann = Annotation('note', '', self)
         fmt.emit_annotation(ann)
